Add check of whether docker image is running when executing.

+++ lib/scripts/docker/exec_kill.sh
@@ -7,4 +7,9 @@ UAS_AT_UCLA_IMAGE=$(docker ps \
   --format "{{.ID}}" \
   --latest)
 
+if [ -z $UAS_AT_UCLA_IMAGE ]
+then
+  exit 1
+fi
+
 docker exec -t $UAS_AT_UCLA_IMAGE sh -c "for f in /tmp/docker-exec-*.pid;do PID=\$(cat \"\$f\");echo \"KILLING \$PID\";kill -15 \$PID > /dev/null 2>&1 || true;rm \$f;done"
